refactor: RacingCarGameRenderer을 추상화하고 추상화에 의존하기 (#8)
* refactor: RacingCarGameRenderer을 SimpleRacingCarGameRenderer로 이름 변경
* refactor: SimpleRacingCarGameRenderer을 RacingCarGameRenderer로 추상화하기
